Refactor package structure: move Dice, Die, GameBoard, and Tile classes to edu.idi.idatt.model
