+++ src/main/java/edu/idi/idatt/model/Die.java
@@ -1,4 +1,4 @@
-package edu.idi.idatt;
+package edu.idi.idatt.model;
 
 import java.util.Random;
