+++ src/main/java/edu/idi/idatt/model/Tile.java
@@ -1,4 +1,4 @@
-package edu.idi.idatt;
+package edu.idi.idatt.model;
 
 import javafx.geometry.Pos;
 import javafx.scene.control.Label;
